requirements.txt
openai
python-dotenv
sounddevice 
soundfile
scipy